refactor: last function which is a conversion
