--- geometric/internal.py
@@ -1,73 +1,0 @@
-"""
-internal.py: Internal coordinate systems
-
-Copyright 2016-2020 Regents of the University of California and the Authors
-
-Authors: Lee-Ping Wang, Chenchen Song
-
-Contributors:
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice,
-this list of conditions and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright notice,
-this list of conditions and the following disclaimer in the documentation
-and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its contributors
-may be used to endorse or promote products derived from this software
-without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
-IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
-INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
-NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-POSSIBILITY OF SUCH DAMAGE.
-"""
-
-from __future__ import division
-
-import numpy as np
-
-from geometric.coordinate_systems.slots import CartesianX, CartesianY, CartesianZ, Distance, \
-    LinearAngle, \
-    TranslationX, \
-    TranslationY, \
-    TranslationZ
-
-from geometric.nifty import bohr2ang
-
-
-## Some vector calculus functions
-
-## End vector calculus functions
-
-def convert_angstroms_degrees(prims, values):
-    """ Convert values of primitive ICs (or differences) from
-    weighted atomic units to Angstroms and degrees. """
-    converted = np.array(values).copy()
-    for ic, c in enumerate(prims):
-        if type(c) in [TranslationX, TranslationY, TranslationZ]:
-            w = 1.0
-        elif hasattr(c, 'w'):
-            w = c.w
-        else:
-            w = 1.0
-        if type(c) in [TranslationX, TranslationY, TranslationZ, CartesianX, CartesianY, CartesianZ, Distance,
-                       LinearAngle]:
-            factor = bohr2ang
-        elif c.isAngular:
-            factor = 180.0 / np.pi
-        converted[ic] /= w
-        converted[ic] *= factor
-    return converted
-
-
